fully separate concerns
